Add an extention point that can deal with container creation
Sometimes, CallListener doesn't quite cut it. There are some use cases where
we'll want to have access to invocation extensions metadata - and possibly,
interact with the container creation process.

The ExtensionInterceptor (should this live in fnext? it has a pretty low-level interface)
adds methods around container creation and deletion - as well as passing more context into
a pair of calls that bracket invocation handling on the live container.

    BeforeCreate(context.Context, drivers.ContainerTask, drivers.Cookie, agent.Call) error
    AfterDestroy(context.Context, drivers.ContainerTask) error

    BeforeCall(context.Context, drivers.ContainerTask, agent.Call, CallExtensions) error
    AfterCall(context.Context, drivers.ContainerTask, agent.Call, CallExtensions) error

(One immediate use for this is to pass principal credentials along with a call.)